Extended the package file list to include the TTF
### Summary ###

The package was publishing all our fonts except the ttf. It had microns.woff, microns.woff2, mirons.eot, microns.svg, but it was missing microns.ttf. This change set add it to the package file list.

It also bumps the version number to v1.0.6, but it'll need the admin to publish that fix to npm.

Resolves #3.